Version 2.2.0.5 Final ================= bear: - Updated the website page. - Fixed a problem where the filtergraph times were not being updated correctly when loading new files. - Fixed a problem with cold starting of Full TS Mux files where the times would change if you changed to a new service. - Added some code to parse the media subtype from the file if its less than 2mb. This is to help with cold starting of timeshift files. - Added some code to help reduce HDD clogging. - Added some code to allow playing the TSFileSink timeshift files from other locations i.e. over networks. - Added a seperate buffer thread to the TSBuffer class, seems to now seek faster. - Fixed a bug where it would not seek properly if the file had a pcr cross over. - Updated the log fileWriter class with a buffer thread. - Fixed a problem with the Demux code so the filter will now work better when changing streams.
git-svn-id: http://svn.nate.deepcreek.org.au/svn/TSFileSource/trunk@115 89fdcf3f-49f1-7541-bbf8-3596d8354d9d

+++ src/FileReader.cpp
@@ -138,7 +138,8 @@ HRESULT FileReader::OpenFile()
 							FILE_SHARE_WRITE,   // Share access
 							NULL,						// Security
 							OPEN_EXISTING,				// Open flags
-							FILE_ATTRIBUTE_NORMAL,		// More flags
+//							(DWORD) 0,
+							(DWORD) FILE_ATTRIBUTE_NORMAL,		// More flags
 //							FILE_ATTRIBUTE_NORMAL |
 //							FILE_FLAG_RANDOM_ACCESS,	// More flags
 //							FILE_FLAG_SEQUENTIAL_SCAN,	// More flags
@@ -163,7 +164,8 @@ HRESULT FileReader::OpenFile()
 			FILE_SHARE_WRITE,   // Share access
 			NULL,      // Security
 			OPEN_EXISTING,    // Open flags
-			FILE_ATTRIBUTE_NORMAL, // More flags
+//			(DWORD) 0,
+			(DWORD) FILE_ATTRIBUTE_NORMAL, // More flags
 //			FILE_FLAG_SEQUENTIAL_SCAN,	// More flags
 //			FILE_ATTRIBUTE_NORMAL |
 //			FILE_FLAG_RANDOM_ACCESS,	// More flags
